% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{abcd_test_data}
\alias{abcd_test_data}
\title{ABCD test data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 582 rows and 13 columns.
}
\usage{
abcd_test_data
}
\description{
ABCD test data
}
\keyword{datasets}
