% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{activity_units}
\alias{activity_units}
\title{Units for economic activities per sector and technology}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 58 rows and 3 columns.
}
\usage{
activity_units
}
\description{
Units for economic activities per sector and technology
}
\keyword{datasets}
