% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_loanbook_exposure_scores.R
\name{aggregate_alignment_loanbook_exposure}
\alias{aggregate_alignment_loanbook_exposure}
\title{Return loan book level aggregation of company alignment metrics by exposure}
\usage{
aggregate_alignment_loanbook_exposure(data, matched, level = c("net", "bo_po"))
}
\arguments{
\item{data}{data.frame. Holds output pf company indicators}

\item{matched}{data.frame. Holds matched and prioritised loan book}

\item{level}{Character. Vector that indicates if the aggregate alignment
metric should be returned based on the net technology deviations (\code{net}) or
disaggregated into buildout and phaseout technologies (\code{bo_po}).}
}
\description{
Return loan book level aggregation of company alignment metrics by exposure
}
