% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_company_alignment_metric.R
\name{calculate_company_aggregate_alignment_sda}
\alias{calculate_company_aggregate_alignment_sda}
\title{Return company level sector alignment metric for each company}
\usage{
calculate_company_aggregate_alignment_sda(
  data,
  scenario_source = "geco_2021",
  scenario = "1.5c"
)
}
\arguments{
\item{data}{data.frame. Holds the PACTA for Banks SDA results on company level.}

\item{scenario_source}{Character. Vector that indicates which scenario_source
to use for reference in the calculation of the alignment metrics. Currently,
the only supported value is \code{"geco_2021"}.}

\item{scenario}{Character. Vector that indicates which scenario to calculate
the alignment metric for. Must be a scenario available from \code{scenario_source}.}
}
\description{
Return company level sector alignment metric for each company
}
