% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_company_alignment_metric.R
\name{calculate_company_aggregate_alignment_tms}
\alias{calculate_company_aggregate_alignment_tms}
\title{Return company level sector alignment metric for each company with
option to disaggregate by buildout / phaseout.}
\usage{
calculate_company_aggregate_alignment_tms(
  data,
  scenario_source = "geco_2021",
  scenario = "1.5c",
  level = c("net", "bo_po")
)
}
\arguments{
\item{data}{data.frame. Holds company-technology deviations based on PACTA
for Banks TMS results. Must have been calculated according to the
increasing/decreasing logic of the CA100+ calculation.}

\item{scenario_source}{Character. Vector that indicates which scenario_source
to use for reference in the calculation of the alignment metrics. Currently,
the only supported value is \code{"geco_2021"}.}

\item{scenario}{Character. Vector that indicates which scenario to calculate
the alignment metric for. Must be a scenario available from \code{scenario_source}.}

\item{level}{Character. Vector that indicates if the aggreagte alignment
metric should be returned based on the net technology deviations (\code{net}) or
disaggregated into buildout and phaseout technologies (\code{bo_po}).}
}
\description{
Return company level sector alignment metric for each company with
option to disaggregate by buildout / phaseout.
}
