% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_company_aggregate_score.R
\name{calculate_company_aggregate_score_sda}
\alias{calculate_company_aggregate_score_sda}
\title{Return company level scenario scores for the main sector of each company}
\usage{
calculate_company_aggregate_score_sda(
  data,
  scenario_emission_intensities,
  scenario_source = "geco_2021",
  scenario = "1.5c",
  aggregate = TRUE
)
}
\arguments{
\item{data}{data.frame. Holds the PACTA for Banks SDA results on company level.}

\item{scenario_emission_intensities}{data frame containing the scenario file with
information on yearly emission intensity levels.}

\item{scenario_source}{Character. Vector that indicates which scenario_source
to use for reference in the calculation of the scores. Currently, the only
supported value is \code{"geco_2021"}.}

\item{scenario}{Character. Vector that indicates which scenario to calculate
the score for. Must be a scenario available from \code{scenario_source}.}

\item{aggregate}{Logical. Indicates whether the indicators should be
calculated for an aggregate of all loan books by different banks in \code{data}
or if they should be calculated individually, based on their
\code{bank_id}. If only one loan book is included, use the default
aggregate == TRUE.}
}
\description{
Return company level scenario scores for the main sector of each company
}
