% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_company_aggregate_score.R
\name{calculate_company_tech_deviation}
\alias{calculate_company_tech_deviation}
\title{Return company level technology deviations for TMS sectors. To be used as
input into calculation of company level aggregate scores for production
trajectory sectors.}
\usage{
calculate_company_tech_deviation(
  data,
  technology_direction,
  scenario_trajectory,
  green_or_brown,
  scenario_source = "geco_2021",
  scenario = "1.5c",
  bridge_tech = c("none", "gascap"),
  aggregate = TRUE
)
}
\arguments{
\item{data}{data.frame. Holds the PACTA for Banks TMS results. Must have been
calculated according to the green/brown logic of the CA100+ calculation
and must return unweighted company level TMSR results.}

\item{technology_direction}{data frame that indicates which technologies are
to be considered phase down technologies versus build out technologies}

\item{scenario_trajectory}{data frame containing the scenario file with
information on yearly tmsr and smsp changes by scenario and region.}

\item{green_or_brown}{data frame. Indicates which technologies are to use
tmsr versus smsp}

\item{scenario_source}{Character. Vector that indicates which scenario_source
to use for reference in the calculation of the scores. Currently, the only
supported value is \code{"geco_2021"}.}

\item{scenario}{Character. Vector that indicates which scenario to calculate
the score for. Must be a scenario available from \code{scenario_source}.}

\item{bridge_tech}{Character. Vector that indicates if a technology is
considered a bridge technology. I.e. if the scenario requires a temporary
build out despite the need for a long term phase down. If so, the alignment
score can be treated differently than for other technologies. Currently,
the only allowed values are (\verb{"none", "gascap"}). Default is \code{"none"} which
means that no special calculations are applied to any technology.}

\item{aggregate}{Logical. Indicates whether the indicators should be
calculated for an aggregate of all loan books by different banks in \code{data}
or if they should be calculated individually, based on their
\code{bank_id}. If only one loan book is included, use the default
aggregate == TRUE.}
}
\description{
Return company level technology deviations for TMS sectors. To be used as
input into calculation of company level aggregate scores for production
trajectory sectors.
}
