% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{capacity_factor_power}
\alias{capacity_factor_power}
\title{Capacity factors for the power technologies based on IEW WEO scenario}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6 rows and 6 columns.
}
\usage{
capacity_factor_power
}
\description{
Capacity factors for the power technologies based on IEW WEO scenario
}
\keyword{datasets}
