% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_benchmark_loanbook.R
\name{create_benchmark_loanbook}
\alias{create_benchmark_loanbook}
\title{Return raw loan book containing the corporate economy benchmark}
\usage{
create_benchmark_loanbook(data)
}
\arguments{
\item{data}{data frame containing the asset-based company data (ABCD) in
PACTA for Banks format.}
}
\description{
Return raw loan book containing the corporate economy benchmark
}
