% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_individual_outputs.R
\name{generate_individual_outputs}
\alias{generate_individual_outputs}
\title{Return data and plots}
\usage{
generate_individual_outputs(
  data,
  matched_loanbook,
  output_directory,
  target_type = c("tms", "sda"),
  bank_id,
  scenario_source,
  target_scenario,
  region = "global",
  sector
)
}
\arguments{
\item{data}{data.frame. Holds the PACTA for Banks results of a set of banks.
Must include a column \code{bank_id} to select individual banks. Can be
either a target market share calculation or an SDA calculation.}

\item{matched_loanbook}{data.frame. Holds the matched loan books of a
set of banks. Must include a column \code{bank_id} to select individual
banks.}

\item{output_directory}{Character. Path to directory where the ouput csv and png
files should be stored.}

\item{target_type}{Character. Must be either \code{tms} or
\code{sda}.}

\item{bank_id}{Character. Must be of length 1.}

\item{scenario_source}{Character. Must be available in the scenario file used
for the analysis.}

\item{target_scenario}{Character. Must be of length 1.}

\item{region}{Character. Must be of length 1.}

\item{sector}{Character. Must be of length 1.}
}
\description{
Return data and plots
}
