% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_energy_sector_split.R
\name{get_energy_sector_split}
\alias{get_energy_sector_split}
\title{Return multi sector energy companies including sector split}
\usage{
get_energy_sector_split(data, start_year)
}
\arguments{
\item{data}{data frame containing the asset-based company data (ABCD) in
PACTA for Banks format.}

\item{start_year}{Numeric vector containing the start year of the analysis.}
}
\description{
Return multi sector energy companies including sector split
}
