% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{loanbook_test_data}
\alias{loanbook_test_data}
\title{Loan book test data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 20 rows and 19 columns.
}
\usage{
loanbook_test_data
}
\description{
Loan book test data
}
\keyword{datasets}
