% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sankey.R
\name{plot_sankey}
\alias{plot_sankey}
\title{Make a sankey plot}
\usage{
plot_sankey(
  data,
  capitalise_node_labels = TRUE,
  save_png_to = NULL,
  png_name = "sankey.png"
)
}
\arguments{
\item{data}{data.frame. Should have the same format as output of
\code{prep_sankey()}}

\item{capitalise_node_labels}{Logical. Flag indicating if node labels should
be converted into better looking capitalised form.}

\item{save_png_to}{Character. Path where the output in png format should be
saved}

\item{png_name}{Character. File name of the output.}
}
\description{
Make a sankey plot
}
\examples{
# TODO
}
