% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scatter.R
\name{plot_scatter}
\alias{plot_scatter}
\title{Plot alignment scatterplot}
\usage{
plot_scatter(
  data,
  sector = NULL,
  scenario_source = NULL,
  scenario = NULL,
  year = NULL,
  region = NULL,
  title = NULL,
  subtitle = NULL,
  alignment_limit = NULL,
  data_level = c("company", "bank")
)
}
\arguments{
\item{data}{data.frame. Should have the same format as output of
\code{prep_scatter()} and contain columns: 'name', 'buildout', phaseout', 'net'.}

\item{sector}{Character. Sector name to be used in the plot title.}

\item{scenario_source}{Character. Scenario source to be used in the plot
caption.}

\item{scenario}{Character. Scenario name to be used in the plot caption.}

\item{year}{Integer. Year of the analysis to be used in the plot caption.}

\item{region}{Character. Region to be used in the plot caption.}

\item{title}{Character. Custom title if different than default.}

\item{subtitle}{Character. Custom subtitle if different than default.}

\item{alignment_limit}{Numeric. Limit to be applied to the x- and y-axis
scales and to alignment values for colouring. By default the maximum
absolute alignment value of is used.}

\item{data_level}{Character. Level of the plotted data. Can be 'bank' or
'company'.}
}
\value{
object of type "ggplot"
}
\description{
Plot alignment scatterplot
}
\examples{
#TODO
}
