% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_sankey.R
\name{prep_sankey}
\alias{prep_sankey}
\title{Prepare data to plot using \code{plot_sankey()}}
\usage{
prep_sankey(
  tms_aggregated,
  sda_aggregated,
  matched_loanbook,
  region_tms,
  region_sda,
  year,
  middle_node = c("sector", "name_abcd")
)
}
\arguments{
\item{tms_aggregated}{data.frame. Holds aggregated alignment scores per
company for tms sectors. Must contain columns: \code{bank_id}, \code{name_abcd},
\code{sector}.}

\item{sda_aggregated}{data.frame. Holds aggregated alignment scores per
company for sda sectors. Must contain columns: \code{bank_id}, \code{name_abcd},
\code{sector}.}

\item{matched_loanbook}{data.frame. Holds the matched loan books of a set of
banks. Must include a column \code{bank_id} and \code{loan_size_outstanding}.}

\item{region_tms}{Character. Region to filter \code{tms_aggregated} data frame on.}

\item{region_sda}{Character. Region to filter \code{sda_aggregated} data frame on.}

\item{middle_node}{Character. Column specifying the middle nodes to be
plotted in sankey plot. Must be present in \code{tms_aggegated} and
\code{sda_aggregated} input data frames.}
}
\value{
data.frame
}
\description{
Prepare data to plot using \code{plot_sankey()}
}
\examples{
# TODO
}
