% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_scatter_animated.R
\name{prep_scatter_animated}
\alias{prep_scatter_animated}
\title{Prepare data to plot animated scatterplot}
\usage{
prep_scatter_animated(
  data_bopo,
  data_net,
  data_level = c("bank", "company"),
  sector,
  region,
  bank_ids_to_plot = NULL
)
}
\arguments{
\item{data_bopo}{data.frame. Data containing buildout and phaseout alignment
values. Must contain columns: 'bank_id', 'year', 'sector', 'region',
'direction' and either 'name_abcd' and 'score' or
'exposure_weighted_net_alignment'.}

\item{data_net}{data.frame. Data containing net alignment values. Must
contain columns: 'bank_id', 'year', 'sector', 'region', 'direction' and
either 'name_abcd' and 'score' or 'exposure_weighted_net_alignment'.}

\item{data_level}{Character. Level of the plotted data. Can be 'bank' or
'company'.}

\item{sector}{Character. Sector to filter data on.}

\item{region}{Character. Region to filter data on.}

\item{bank_ids_to_plot}{Character vector. Bank ids to filter on.}
}
\value{
data.frame
}
\description{
Prepare data to plot animated scatterplot
}
\examples{
#TODO
}
