% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{primary_energy_efficiency}
\alias{primary_energy_efficiency}
\title{Primary energy efficiency factors by power technology}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6 rows and 3 columns.
}
\usage{
primary_energy_efficiency
}
\description{
Primary energy efficiency factors by power technology
}
\keyword{datasets}
