% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{unit_conversion}
\alias{unit_conversion}
\title{Units for economic activities per sector and technology}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3 rows and 3 columns.
}
\usage{
unit_conversion
}
\description{
Units for economic activities per sector and technology
}
\keyword{datasets}
