% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_company_alignment_metric.R
\name{calculate_company_tech_deviation}
\alias{calculate_company_tech_deviation}
\title{Return company level technology deviations for TMS sectors. To be used as
input into calculation of company level aggregate alignment metrics for
production trajectory sectors.}
\usage{
calculate_company_tech_deviation(
  data,
  technology_direction,
  scenario_source = "geco_2021",
  scenario = "1.5c",
  bridge_tech = c("none", "gascap"),
  time_frame = 5L
)
}
\arguments{
\item{data}{data.frame. Holds the PACTA for Banks TMS results. Must have been
calculated according to the increasing/decreasing logic of the CA100+ calculation
and must return unweighted company level TMSR results.}

\item{technology_direction}{data frame that indicates which technologies are
to be considered phase down technologies versus build out technologies}

\item{scenario_source}{Character. Vector that indicates which scenario_source
to use for reference in the calculation of the alignment metrics. Currently,
the only supported value is \code{"geco_2021"}.}

\item{scenario}{Character. Vector that indicates which scenario to calculate
the alignment metric for. Must be a scenario available from \code{scenario_source}.}

\item{bridge_tech}{Character. Vector that indicates if a technology is
considered a bridge technology. I.e. if the scenario requires a temporary
build out despite the need for a long term phase down. If so, the alignment
metric can be treated differently than for other technologies. Currently,
the only allowed values are (\verb{"none", "gascap"}). Default is \code{"none"} which
means that no special calculations are applied to any technology.}

\item{time_frame}{Integer of length one. The number of forward looking years
that should be considered in the analysis. Standard \code{time_frame} in PACTA
is five years.}
}
\description{
Return company level technology deviations for TMS sectors. To be used as
input into calculation of company level aggregate alignment metrics for
production trajectory sectors.
}
