% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_energy_sector_split.R
\name{get_energy_sector_split}
\alias{get_energy_sector_split}
\title{Return multi sector energy companies including sector split}
\usage{
get_energy_sector_split(
  data,
  start_year,
  primary_energy_efficiency,
  unit_conversion
)
}
\arguments{
\item{data}{data frame containing the asset-based company data (ABCD) in
PACTA for Banks format.}

\item{start_year}{Numeric vector containing the start year of the analysis.}

\item{primary_energy_efficiency}{data frame containing efficiency factors for
power technologies that can be used to back calculated the primary energy
content implied in a unit of power capacity.}

\item{unit_conversion}{data frame containing conversion factors for all
energy sectors covered in PACTA so that the metrics can be transformed to a
single common energy unit.}
}
\description{
Return multi sector energy companies including sector split
}
