% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_sankey.R
\name{prep_sankey}
\alias{prep_sankey}
\title{Prepare data to plot using \code{plot_sankey()}}
\usage{
prep_sankey(
  data_alignment,
  matched_loanbook,
  region,
  year,
  middle_node,
  middle_node2 = NULL
)
}
\arguments{
\item{data_alignment}{data.frame. Holds aggregated alignment metrics per
company for tms sectors. Must contain columns: \code{group_id}, \code{name_abcd},
\code{sector}.}

\item{matched_loanbook}{data.frame. Holds the matched loan books of a set of
groups. Must include a column \code{group_id} and \code{loan_size_outstanding}.}

\item{region}{Character. Region to filter \code{data_alignment} data frame on.}

\item{year}{Integer. Year on which \code{data_alignment} should be filtered.}

\item{middle_node}{Character. Column specifying the middle nodes to be
plotted in sankey plot. Must be present in \code{data_alignment}.}

\item{middle_node2}{Character. Column specifying the middle nodes to be
plotted in sankey plot. Must be present in \code{data_alignment}.}
}
\value{
data.frame
}
\description{
Prepare data to plot using \code{plot_sankey()}
}
\examples{
# TODO
}
