% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_timeline.R
\name{prep_timeline}
\alias{prep_timeline}
\title{Prepare data to plot timeline}
\usage{
prep_timeline(data, sector, bank_ids_to_plot)
}
\arguments{
\item{data}{data.frame. Must contain columns: 'direction', 'year',
'exposure_weighted_net_alignment', 'bank_id', 'sector'.}

\item{sector}{Character. Sector to filter data on.}

\item{bank_ids_to_plot}{Character vector. Bank ids to filter on.}
}
\value{
data.frame
}
\description{
Prepare data to plot timeline
}
\examples{
#TODO
}
